tests/wpt/tests/css/css-anchor-position/anchor-getComputedStyle-002.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Tests getComputedStyle() resolving anchor() in fragmentation context</title>
<link rel="help" href="https://drafts4.csswg.org/css-anchor-position-1">
<link rel="help" href="https://drafts.csswg.org/cssom/#resolved-value">
<link rel="author" href="mailto:xiaochengh@chromium.org">
<link rel="stylesheet" href="/fonts/ahem.css">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
body {
  margin: 0;
}

.cb {
  position: relative;
  background: lightgray;
}

.anchor {
  anchor-name: --a;
  background: orange;
}

.target {
  position: absolute;
  left: anchor(--a left);
  right: anchor(--a right);
  top: anchor(--a top);
  bottom: anchor(--a bottom);
  background: lime;
  opacity: 0.5;
}
</style>

<!-- anchor is fragmented in second and third columns -->
<div class="multicol" id="test1">
  <div class="cb">
    <div class="spacer"></div>
    <div class="anchor"></div>
    <div class="target"></div>
  </div>
</div>
<style>
#test1.multicol {
  column-count: 3;
  column-width: 100px;
  column-gap: 10px;
  width: 320px;
  height: 100px;
}

#test1 .cb {
  width: 100px;
  height: 300px;
}

#test1 .spacer {
  height: 175px;
}

#test1 .anchor {
  margin-left: 25px;
  width: 50px;
  height: 50px;
}
</style>
<script>
test(() => {
  const target = test1.querySelector('.target');
  const style = getComputedStyle(target);
  assert_equals(style.left, '25px');
  assert_equals(style.right, '-85px');
  assert_equals(style.top, '100px');
  assert_equals(style.bottom, '100px');
}, 'getComputedStyle() with fragmented containing block in multicolumn layout');
</script>


<div id="test2" style="font: 20px/1 Ahem; width: 11em">
  Lorem
  <span class="cb">
    ipsum <span class="anchor">dolor</span> sit
    <span class="target"></span>
  </span>
  amet.<br>

  Lorem
  <span class="cb">
    ipsum dolor <span class="anchor">sit</span>
    <span class="target"></span>
  </span>
  amet.<br>
</div>
<script>
test(() => {
  const targets = test2.querySelectorAll('.target');

  const style1 = getComputedStyle(targets[0]);
  assert_equals(style1.top, '20px');
  assert_equals(style1.bottom, '0px');
  assert_equals(style1.left, '-120px');
  assert_equals(style1.right, '80px');

  const style2 = getComputedStyle(targets[1]);
  assert_equals(style2.top, '20px');
  assert_equals(style2.bottom, '0px');
  assert_equals(style2.left, '0px');
  assert_equals(style2.right, '0px');
}, 'getComputedStyle() with fragmented containing block in inline layout');
</script>
