third_party/blink/web_tests/platform/linux-chrome/external/wpt/css/css-anchor-position/anchor-getComputedStyle-002-expected.txt

This is a testharness.js-based test.
[PASS] getComputedStyle() with fragmented containing block in multicolumn layout
[FAIL] getComputedStyle() with fragmented containing block in inline layout
  assert_equals: expected "20px" but got "0px"
Harness: the test ran to completion.
